Add sprig library for template functions.
Fixes #2737
Alternative to #2738

Removes duplicates from the funcmap that exist already in sprig.
Note that this does leave the `hasPrefix` function overriding the one in
sprig. This is because the sprig function expects arguments of the form
`(prefix, string)` and the existing go-swagger function expects
`(string, prefix)`.

As to not cause a breaking change to the existing templating this was
left as-is.

Signed-off-by: Kevin Barbour <kevinbarbourd@gmail.com>