fix(mangling): overlapping pluralized initialisms trigger panic
Fuzz test detected an interesting fault in the initialisms matcher:
string "TTLss" triggered a panic.

The pluralization of "TTLs", together with a possible match with "TLS"
caused a panic.

This PR added a fuzz test for ToGoName, fixes the issue and adds the
detected string to the fuzzing corpus (testdata).

Signed-off-by: Frédéric BIDON <fredbi@yahoo.com>